make torque control work on the robot with the gripper mounted, remove collision models in urdf for a faster simulation, and need to improve the gravity compensation part by using Jacobian method.
