+++ drake/examples/kuka_iiwa_arm/Robotiq_Command.py
@@ -7,11 +7,6 @@ import CModelSimpleController
 from robo_lcm import lcmt_robotiq_output as outputMsg
 
 class Robotiq_Command(object):
-    '''
-    Displays simulation time and frequency in the status bar
-    using information from the IIWA_STATUS lcm channel.
-    '''
-
     def __init__(self):
         self.command = outputMsg()
         self.lc = lcm.LCM()
@@ -19,28 +14,17 @@ class Robotiq_Command(object):
         self.command = outputMsg()
 
     def gripperConnect(self):
-      #CModelRtuNode.mainLoop("/dev/ttyUSB0")
-      #os.system("CModelRtuNode.py /dev/ttyUSB0")
       CModelSimpleController.publisher()
       pass
     def gripperOpen(self):
-      #sendGripperCommand(0, speed=255, force=150)
       self.command = CModelSimpleController.genCommand('o',self.command)
       self.lc.publish("command_topic",self.command.encode())
-      #pass
     def gripperClose(self):
-      #sendGripperCommand(255, speed=255, force=150)
       self.command = CModelSimpleController.genCommand('c',self.command)
       self.lc.publish("command_topic",self.command.encode())
-      #pass
     def gripperReset(self):
-      #sendGripperCommand(0, speed=255, force=150)
-      #sendGripperCommand('r')
       self.command = CModelSimpleController.genCommand('r',self.command)
       self.lc.publish("command_topic",self.command.encode())
-      #pass
     def gripperActivate(self):
-      #sendGripperCommand(0, speed=255, force=150)
-      #sendGripperCommand('a')
       self.command = CModelSimpleController.genCommand('a',self.command)
       self.lc.publish("command_topic",self.command.encode())
