+++ drake/examples/kuka_iiwa_arm/kuka_simulation.cc
@@ -165,7 +165,7 @@ class SimulatedKuka : public systems::Diagram<T> {
       auto rigid_body_tree = std::make_unique<RigidBodyTree<double>>();
       drake::parsers::urdf::AddModelInstanceFromUrdfFile(
           drake::GetDrakePath() +
-          "/examples/kuka_iiwa_arm/urdf/iiwa14_fixed_gripper.urdf", //iiwa14_fixed_gripper
+          "/examples/kuka_iiwa_arm/urdf/iiwa14_fixed_gripper.urdf", 
           drake::multibody::joints::kFixed,
           nullptr /* weld to frame */, rigid_body_tree.get());
