+++ drake/examples/kuka_iiwa_arm/urdf/iiwa14.urdf
@@ -59,13 +59,13 @@
       </geometry>
       <material name="Grey"/>
     </visual>
-    <collision>
+<!--     <collision>
       <origin rpy="0 0 0" xyz="-0.015 0 0.07"/>
       <geometry>
           <cylinder radius="0.139" length="0.17"/>
       </geometry>
       <material name="Grey"/>
-    </collision>
+    </collision> -->
     <self_collision_checking>
       <origin rpy="0 0 0" xyz="0 0 0"/>
       <geometry>
@@ -97,13 +97,13 @@
       </geometry>
       <material name="Grey"/>
     </visual>
-    <collision>
+<!--     <collision>
       <origin rpy="0 0 0" xyz="0 -0.014 0.14"/>
       <geometry>
           <cylinder radius="0.101" length="0.29"/>
       </geometry>
       <material name="Grey"/>
-    </collision>
+    </collision> -->
   </link>
   <!-- joint between link_1 and link_2 -->
   <joint name="iiwa_joint_2" type="revolute">
@@ -129,13 +129,13 @@
       </geometry>
       <material name="Orange"/>
     </visual>
-    <collision>
+<!--     <collision>
       <origin rpy="1.57079632679 0 0" xyz="0 0.056 0.035"/>
       <geometry>
           <cylinder radius="0.101" length="0.29"/>
       </geometry>
       <material name="Orange"/>
-    </collision>
+    </collision> -->
   </link>
   <!-- joint between link_2 and link_3 -->
   <joint name="iiwa_joint_3" type="revolute">
@@ -161,13 +161,13 @@
       </geometry>
       <material name="Grey"/>
     </visual>
-    <collision>
+<!--     <collision>
       <origin rpy="0 0 0" xyz="0 0.025 0.137"/>
       <geometry>
           <cylinder radius="0.092" length="0.29"/>
       </geometry>
       <material name="Grey"/>
-    </collision>
+    </collision> -->
   </link>
   <!-- joint between link_3 and link_4 -->
   <joint name="iiwa_joint_4" type="revolute">
@@ -193,13 +193,13 @@
       </geometry>
       <material name="Orange"/>
     </visual>
-    <collision>
+<!--     <collision>
       <origin rpy="1.57079632679 0 0" xyz="0 0.056 0.025"/>
       <geometry>
           <cylinder radius="0.092" length="0.25"/>
       </geometry>
       <material name="Orange"/>
-    </collision>
+    </collision> -->
   </link>
   <!-- joint between link_4 and link_5 -->
   <joint name="iiwa_joint_5" type="revolute">
@@ -225,13 +225,13 @@
       </geometry>
       <material name="Grey"/>
     </visual>
-    <collision>
+<!--     <collision>
       <origin rpy="-0.3 0 0" xyz="0 0.03 0.13"/>
       <geometry>
           <cylinder radius="0.07" length="0.29"/>
       </geometry>
       <material name="Grey"/>
-    </collision>
+    </collision> -->
   </link>
   <!-- joint between link_5 and link_6 -->
   <joint name="iiwa_joint_6" type="revolute">
@@ -257,13 +257,13 @@
       </geometry>
       <material name="Orange"/>
     </visual>
-    <collision>
+<!--     <collision>
       <origin rpy="0 0 0" xyz="0 0 0"/>
       <geometry>
         <mesh filename="meshes/collision/link_6.obj"/>
       </geometry>
       <material name="Orange"/>
-    </collision>
+    </collision> -->
   </link>
   <!-- joint between link_6 and link_7 -->
   <joint name="iiwa_joint_7" type="revolute">
