--- drake/examples/kuka_iiwa_arm/urdf/meshes/stl_OBJ.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
- # echo of all files in a directory
-
-for file in *.stl
-do
-  name=${file%%[.]*}
-  meshlabserver -i $file -o $name'.obj' -om vn
-  meshlabserver -i $file -o $name'_chull.obj' -om vn -s chull.mlx
-  #echo $name'.obj' 
-done
