+++ drake/lcmtypes/CMakeLists.txt
@@ -42,8 +42,6 @@ lcm_wrap_types(
   lcmt_scope_data.lcm
   lcmt_schunk_command.lcm
   lcmt_schunk_status.lcm
-  lcmt_schunk_wsg_command.lcm
-  lcmt_schunk_wsg_status.lcm
   lcmt_robotiq_command.lcm
   lcmt_robotiq_status.lcm
   lcmt_simple_car_config_t.lcm
