--- drake/lcmtypes/lcmt_schunk_wsg_command.lcm
@@ -1,9 +1,0 @@
-package drake;
-
-struct lcmt_schunk_wsg_command
-{
-  int64_t utime;  //< The timestamp in microseconds.
-
-  double target_position_mm;
-  double force;
-}
