--- drake/lcmtypes/lcmt_schunk_wsg_status.lcm
@@ -1,10 +1,0 @@
-package drake;
-
-struct lcmt_schunk_wsg_status
-{
-  int64_t utime;  //< The timestamp in microseconds.
-
-  double actual_position_mm;
-  double actual_force;
-  double actual_speed_mm_per_s;
-}
